main_requirements.txt
datasets==3.0.1
transformers==4.45.1
torch==2.4.1
torchvision==0.19.1
huggingface-hub==0.25.1
numpy==2.1.1
pandas==2.2.3
pillow==10.4.0
matplotlib==3.9.2
ipython==8.27.0
tqdm==4.66.5
transformers==4.45.1
git+https://github.com/openai/CLIP.git
nibabel==5.2.1
batchgenerators==0.25
requests==2.32.3
pyqt5==5.15.11
